[MOV] im_livechat: move WebsiteLivechat outside of public_livechat.js
Task-2212347
